requirements.txt
PyYAML==6.0.2
openpyxl==3.1.5
google-api-python-client==2.156.0
google-auth-oauthlib==1.2.1
oauth2client==4.1.3
emoji==2.14.0
python-telegram-bot==21.9
telethon==1.38.1
cython==0.29.33
numpy==2.2.0
pandas==2.2.3
